pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.2",
    "dnspython>=2.7.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pyopenssl>=25.0.0",
    "python-dotenv>=1.1.0",
    "python-whois>=0.9.5",
    "redis>=6.0.0",
    "requests>=2.32.3",
    "sslyze>=6.1.0",
    "trafilatura>=2.0.0",
]
